live/requirements.txt
kafka-python==2.0.2
selenium==4.17.2
beautifulsoup4==4.10.0
discord_webhook==1.3.1